+++ typescript/cli/src/deploy/warp.ts
@@ -3,7 +3,11 @@ import { stringify as yamlStringify } from 'yaml';
 
 import { buildArtifact as coreBuildArtifact } from '@hyperlane-xyz/core/buildArtifact.js';
 import { AltVMWarpModule } from '@hyperlane-xyz/deploy-sdk';
-import { GasAction, ProtocolType } from '@hyperlane-xyz/provider-sdk';
+import {
+  AltVMJsonRpcTxSubmitter,
+  GasAction,
+  ProtocolType,
+} from '@hyperlane-xyz/provider-sdk';
 import {
   AddWarpRouteConfigOptions,
   BaseRegistry,
@@ -58,7 +62,6 @@ import {
 } from '@hyperlane-xyz/utils';
 
 import { TypedAnnotatedTransaction } from '../../../sdk/dist/providers/ProviderType.js';
-import { createAltVMSubmitterFactories } from '../context/altvm.js';
 import { requestAndSaveApiKeys } from '../context/context.js';
 import { WriteCommandContext } from '../context/types.js';
 import {
@@ -70,8 +73,10 @@ import {
   warnYellow,
 } from '../logger.js';
 import { WarpSendLogs } from '../send/transfer.js';
+import { AltVMFileSubmitter } from '../submitters/AltVMFileSubmitter.js';
 import { EV5FileSubmitter } from '../submitters/EV5FileSubmitter.js';
 import {
+  CustomTxSubmitterType,
   ExtendedChainSubmissionStrategy,
   ExtendedChainSubmissionStrategySchema,
   ExtendedSubmissionStrategy,
@@ -1015,19 +1020,40 @@ export async function getSubmitterByStrategy<T extends ProtocolType>({
       : defaultSubmitter;
 
   const strategyToUse = submissionStrategy ?? defaultSubmitter;
+  const protocol = multiProvider.getProtocol(chain);
+
+  const additionalSubmitterFactories: any = {
+    [ProtocolType.Ethereum]: {
+      file: (_multiProvider: MultiProvider, metadata: any) => {
+        return new EV5FileSubmitter(metadata);
+      },
+    },
+  };
+
+  // Only add non-Ethereum protocol factories if we have an alt VM signer
+  if (protocol !== ProtocolType.Ethereum) {
+    const signer = mustGet(altVmSigners, chain);
+    additionalSubmitterFactories[protocol] = {
+      jsonRpc: () => {
+        return new AltVMJsonRpcTxSubmitter(signer, {
+          chain: chain,
+        });
+      },
+      [CustomTxSubmitterType.FILE]: (
+        _multiProvider: MultiProvider,
+        metadata: any,
+      ) => {
+        return new AltVMFileSubmitter(signer, metadata);
+      },
+    };
+  }
+
   return {
     submitter: await getSubmitterBuilder<T>({
       submissionStrategy: strategyToUse as SubmissionStrategy, // TODO: fix this
       multiProvider,
       coreAddressesByChain: await registry.getAddresses(),
-      additionalSubmitterFactories: {
-        [ProtocolType.Ethereum]: {
-          file: (_multiProvider: MultiProvider, metadata: any) => {
-            return new EV5FileSubmitter(metadata);
-          },
-        },
-        ...createAltVMSubmitterFactories(multiProvider, altVmSigners, chain),
-      },
+      additionalSubmitterFactories,
     }),
     config: submissionStrategy,
   };