+++ starknet/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @hyperlane-xyz/starknet-core
 
+## 19.13.0
+
 ## 19.12.0
 
 ## 19.11.0